shift all crlb calculations to precision.py
